follow up on dict issue
Yes this was issue. I was going into the single dicts and not the dict of dicts.

the fix was to add a [2] in front of the get - this would get info from dict #2

so now i just need to put this in a loop

also, I totally forgot you can open excel data in vs code! no more opening the excel file!